[Security Solution][Timeline] refactor flyout folder
