+++ x-pack/plugins/security_solution/public/management/cypress/e2e/response_actions/alerts_response_console.cy.ts
@@ -104,7 +104,7 @@ describe(
 
       getAlertsTableRows().should('have.length.greaterThan', 0);
       openInvestigateInTimelineView();
-      cy.getByTestSubj('timeline-flyout').within(() => {
+      cy.getByTestSubj('timeline-container').within(() => {
         openAlertDetailsView();
       });
       openResponderFromEndpointAlertDetails();
