--- x-pack/plugins/security_solution/public/timelines/components/flyout/pane/translations.ts
@@ -1,15 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { i18n } from '@kbn/i18n';
-
-export const TIMELINE_DESCRIPTION = i18n.translate(
-  'xpack.securitySolution.timeline.flyout.pane.timelinePropertiesAriaLabel',
-  {
-    defaultMessage: 'Timeline Properties',
-  }
-);
