+++ x-pack/plugins/security_solution/public/timelines/components/timeline/index.tsx
@@ -18,7 +18,7 @@ import { timelineDefaults } from '../../store/defaults';
 import { defaultHeaders } from './body/column_headers/default_headers';
 import type { CellValueElementProps } from './cell_rendering';
 import { SourcererScopeName } from '../../../common/store/sourcerer/model';
-import { FlyoutHeaderPanel } from '../flyout/header';
+import { TimelineModalHeader } from '../modal/header';
 import type { TimelineId, RowRenderer, TimelineTabs } from '../../../../common/types/timeline';
 import { TimelineType } from '../../../../common/api/timeline';
 import { useDeepEqualSelector, useShallowEqualSelector } from '../../../common/hooks/use_selector';
@@ -37,6 +37,12 @@ const TimelineTemplateBadge = styled.div`
   color: #fff;
   padding: 10px 15px;
   font-size: 0.8em;
+`;
+
+const TimelineBody = styled.div`
+  height: 100%;
+  display: flex;
+  flex-direction: column;
 `;
 
 export const TimelineContext = createContext<{ timelineId: string | null }>({ timelineId: null });
@@ -216,7 +222,7 @@ const StatefulTimelineComponent: React.FC<Props> = ({
         ref={containerElement}
       >
         <TimelineSavingProgress timelineId={timelineId} />
-        <div className="timeline-body" data-test-subj="timeline-body">
+        <TimelineBody data-test-subj="timeline-body">
           {timelineType === TimelineType.template && (
             <TimelineTemplateBadge className="timeline-template-badge">
               {i18n.TIMELINE_TEMPLATE}
@@ -227,7 +233,7 @@ const StatefulTimelineComponent: React.FC<Props> = ({
             $isVisible={!timelineFullScreen}
             data-test-subj="timeline-hide-show-container"
           >
-            <FlyoutHeaderPanel timelineId={timelineId} />
+            <TimelineModalHeader timelineId={timelineId} />
           </HideShowContainer>
 
           <TabsContent
@@ -240,7 +246,7 @@ const StatefulTimelineComponent: React.FC<Props> = ({
             timelineDescription={description}
             timelineFullScreen={timelineFullScreen}
           />
-        </div>
+        </TimelineBody>
       </TimelineContainer>
       {showTimelineTour ? (
         <TimelineTour
