+++ x-pack/test/security_solution_cypress/cypress/tasks/security_main.ts
@@ -6,8 +6,9 @@
  */
 
 import { recurse } from 'cypress-recurse';
-import { CLOSE_TIMELINE_BUTTON, TIMELINE_BOTTOM_BAR_TOGGLE_BUTTON } from '../screens/security_main';
+import { TIMELINE_BOTTOM_BAR_TOGGLE_BUTTON } from '../screens/security_main';
 import {
+  CLOSE_TIMELINE_BTN,
   TIMELINE_EXIT_FULL_SCREEN_BUTTON,
   TIMELINE_FULL_SCREEN_BUTTON,
   TIMELINE_WRAPPER,
@@ -20,12 +21,12 @@ export const openTimelineUsingToggle = () => {
       return cy.get(TIMELINE_WRAPPER);
     },
     // Retry if somehow the timeline wrapper is still hidden
-    ($timelineWrapper) => !$timelineWrapper.hasClass('timeline-wrapper--hidden')
+    ($timelineWrapper) => !$timelineWrapper.hasClass('timeline-portal-overlay-mask--hidden')
   );
 };
 
 export const closeTimelineUsingCloseButton = () => {
-  cy.get(CLOSE_TIMELINE_BUTTON).filter(':visible').click();
+  cy.get(CLOSE_TIMELINE_BTN).filter(':visible').click();
 };
 
 export const enterFullScreenMode = () => {
