+++ x-pack/test/security_solution_ftr/page_objects/timeline/index.ts
@@ -10,7 +10,7 @@ import { DATE_RANGE_OPTION_TO_TEST_SUBJ_MAP } from '@kbn/security-solution-plugi
 import { FtrService } from '../../../functional/ftr_provider_context';
 
 const TIMELINE_BOTTOM_BAR_CONTAINER_TEST_SUBJ = 'timeline-bottom-bar';
-const TIMELINE_CLOSE_BUTTON_TEST_SUBJ = 'close-timeline';
+const TIMELINE_CLOSE_BUTTON_TEST_SUBJ = 'timeline-modal-header-close-button';
 const TIMELINE_MODAL_PAGE_TEST_SUBJ = 'timeline';
 const TIMELINE_TAB_QUERY_TEST_SUBJ = 'timeline-tab-content-query';
